Implement new glfw.ContextWatcher interface.
Call gl.Init on Windows platform after a context is made current. This
should be correct, especially if the new context is significantly
different from the last (e.g., using a different vendor video card on a
computer with multiple video cards). It may be relatively slow (60-80
ms), so if context switching needs to be done often (not just at
initial window creation), a better solution may be needed and an issue
should be open.

Fixes #6.