Call gl/v2.1/gl.Init() at most once.
The motivation is that the vast majority of Go OpenGL apps behave in
this way, rather than trying to be more correct and initialize GL
bindings for every different context. This should work in most cases.
If there's ever a reproducible situation where this optimization starts
to cause issues, then it can be addressed by either reverting to a
slower solution (initialize GL bindings when making current contexts
that are new/different), or even caching and maintaining a GL binding
per context. This is additional code that should be done once there's a
need.

Resolves #7.