Add GetBoundFramebuffer method.
This method allows to get the currently bound frame buffer on all
platforms. There was a problem where on OpenGL it expects an integer
ID to represent the framebuffer, but on WebGL it expects a WebGLFramebuffer
*js.Object. Adding this method appears to be the only way to expose
the currently bound framebuffer via a uniform API.

This API is somewhat experimental since it falls outside of OpenGL spec,
and may be removed/updated/replaced, if we find a better way of achieving
this goal.

Resolves #20.
Closes #21.