test: support darwin/arm64 and silence OpenGL deprecation warning
Remove the amd64 build constraint from context_darwin_amd64.go,
the test code works on darwin/arm64 too.

The OpenGL API has been deprecated on macOS as of version 10.14.¹
Building github.com/goxjs/gl/test on macOS 11.1 produces warnings:

	context_darwin.go:30:2: warning: 'CGLChoosePixelFormat' is deprecated: first deprecated in macOS 10.14 - OpenGL API deprecated. (Define GL_SILENCE_DEPRECATION to silence these warnings) [-Wdeprecated-declarations]
	/Applications/Xcode.app/Contents/Developer/Platforms/MacOSX.platform/Developer/SDKs/MacOSX.sdk/System/Library/Frameworks/OpenGL.framework/Headers/OpenGL.h:31:17: note: 'CGLChoosePixelFormat' has been explicitly marked deprecated here
	context_darwin.go:31:2: warning: 'CGLCreateContext' is deprecated: first deprecated in macOS 10.14 - OpenGL API deprecated. (Define GL_SILENCE_DEPRECATION to silence these warnings) [-Wdeprecated-declarations]
	...

The package is named gl, so it would be too drastic of a change
for it to stop using OpenGL. Also, OpenGL is not deprecated on all
other platforms that package gl supports.

Despite being deprecated, OpenGL is not removed from macOS 11.1 at this
time. It's not helpful for the warnings to be shown to users who are
building and using the package on macOS, so, simply acknowledge the
deprecation notice by defining GL_SILENCE_DEPRECATION and silence the
warnings.

¹ https://developer.apple.com/documentation/macos_release_notes/macos_mojave_10_14_release_notes?language=objc#3035786